A bit of clanup and some stylistic changes.
When nightmode changed there was a request to do a new update, I
don't think that's needed as all the color changes happens in the
current update after that point anywhere. That also makes it
uneccessary to set lastUpdate to null.

Changed some formatting to not have { on new line.